+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Feature) Expose core.PodSecurityContext Sysctl options
 - (Bugfix) Skip Collection check for missing Database
 - (Feature) Abort resignation of leadership when DB server is restared
+- (Feature) Numactl Options
 
 ## [1.2.31](https://github.com/arangodb/kube-arangodb/tree/1.2.31) (2023-07-14)
 - (Improvement) Block traffic on the services if there is more than 1 active leader in ActiveFailover mode